+++ mrv2/docs/HISTORY.md
@@ -5,6 +5,8 @@ v1.1.4
 - Prevented a macOS python plug-in issue with Python 3.11.
 - Fixed Wayland support on Ubuntu 22.04.3 LTS and later.
   Still pending is how to support Clip dragging.
+- Added Windows' stack traces with line numbers on debug builds.
+- Allowed pyFLTK be able to compile in Debug.
 
 
 v1.1.3